Split authorities in the list on the columns - title and type + inor fixes

+++ Kitodo/src/main/java/org/kitodo/production/forms/AuthorityForm.java
@@ -32,6 +32,8 @@ public class AuthorityForm extends BaseForm {
     private static final long serialVersionUID = 3541160917458068675L;
     private static final Logger logger = LogManager.getLogger(RoleForm.class);
     private Authority authority = new Authority();
+    private String title;
+    private String type;
     private final String authorityListPath = MessageFormat.format(REDIRECT_PATH, "users");
     private final String authorityEditPath = MessageFormat.format(REDIRECT_PATH, "authorityEdit");
 
@@ -41,6 +43,44 @@ public class AuthorityForm extends BaseForm {
     public AuthorityForm() {
         super();
         super.setLazyDTOModel(new LazyDTOModel(ServiceManager.getAuthorityService()));
+    }
+
+    /**
+     * Get title.
+     *
+     * @return value of title
+     */
+    public String getTitle() {
+        return title;
+    }
+
+    /**
+     * Set title.
+     *
+     * @param title
+     *            as java.lang.String
+     */
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    /**
+     * Get type.
+     *
+     * @return value of type
+     */
+    public String getType() {
+        return type;
+    }
+
+    /**
+     * Set type.
+     *
+     * @param type
+     *            as java.lang.String
+     */
+    public void setType(String type) {
+        this.type = type;
     }
 
     /**
@@ -60,6 +100,7 @@ public class AuthorityForm extends BaseForm {
      */
     public String save() {
         try {
+            this.authority.setTitle(this.title + "_" + this.type);
             ServiceManager.getAuthorityService().saveToDatabase(this.authority);
             return authorityListPath;
         } catch (DAOException e) {
@@ -74,7 +115,10 @@ public class AuthorityForm extends BaseForm {
      */
     public void delete() {
         try {
-
+            if (!this.authority.getRoles().isEmpty()) {
+                Helper.setErrorMessage("authorityAssignedError");
+                return;
+            }
             ServiceManager.getAuthorityService().removeFromDatabase(this.authority);
         } catch (DAOException e) {
             Helper.setErrorMessage(ERROR_DELETING, new Object[] {ObjectType.AUTHORITY.getTranslationSingular() },
@@ -83,22 +127,31 @@ public class AuthorityForm extends BaseForm {
     }
 
     /**
-     * Method being used as viewAction for role edit form. Selectable clients and
-     * projects are initialized as well.
+     * Method being used as viewAction for authority edit form.
      *
      * @param id
-     *            ID of the role to load
+     *            ID of the authority to load
      */
     public void load(int id) {
-        try {
-            if (!Objects.equals(id, 0)) {
-                setAuthority(ServiceManager.getAuthorityService().getById(id));
-            }
-        } catch (DAOException e) {
-            Helper.setErrorMessage(ERROR_LOADING_ONE, new Object[] {ObjectType.AUTHORITY.getTranslationSingular(), id },
-                logger, e);
+        if (!Objects.equals(id, 0)) {
+            setAuthorityById(id);
         }
         setSaveDisabled(true);
+    }
+
+    /**
+     * Set authority by id.
+     *
+     * @param id
+     *            ID of authority to set
+     */
+    public void setAuthorityById(int id) {
+        try {
+            setAuthority(ServiceManager.getAuthorityService().getById(id));
+        } catch (DAOException e) {
+            Helper.setErrorMessage(ERROR_LOADING_ONE, new Object[] {ObjectType.ROLE.getTranslationSingular(), id },
+                logger, e);
+        }
     }
 
     /**
@@ -109,6 +162,8 @@ public class AuthorityForm extends BaseForm {
      */
     public void setAuthority(Authority authority) {
         this.authority = authority;
+        this.title = this.authority.getTitleWithoutSuffix();
+        this.type = this.authority.getType();
     }
 
     /**
